add ECR repo for finch rootfs image (#355)
Add an ECR repo to Finch infrastructure. this repository is where the
image used as the basis for the Finch VM rootfs will live.

to support Finch on Windows, we will be using WSL2.
WSL2 differs from the Finch on Mac qemu/vz implementation in that WSL2
provides a kernel and requires just a root filesystem to boot a custom
distribution. We will be maintaining the root filesystem as an image
in this ECR repository. Subsequent PRs to finch-core will include the
Dockerfile, actions for exporting the image as an archive and uploading
to the Finch dependencies S3 bucket.

For now, the ECR repository is private to limit access controls to the
GitHub action responsible for interacting with the image. However if in
the future that proves to be cumbersome for developers, we can make the
repository public for image pull and limit push permissions to the action.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>