feat: add rootfs image scanning notifications
Amazon ECR offers Enhanced Image Scanning, which we have enabled for the
rootfs image used for Finch on Windows. This change implements a stack
to automatically notify Finch developers of scan results when the
finding is HIGH or CRITICAL in severity (aka CVE score >= 7.0).

An SNS topic is created for the notification, and a Lambda function is
wired to EventBridge to listen for events from AWS Inspector of those
severities.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>