fix: give Administrator log on as service rights
The GitHub Actions Runner service is configured to be logged in as the
Administrator user such that the service has correct permissions to
interact with WSL. Previously, when the service started on instance
reboot, it would fail because the Administrator user is not in the set
of users granted permissions to "Log on As Service". Fix by using
'Carbon' powershell module (https://get-carbon.org/Grant-Privilege.html)
that grants the Administrator user proper rights.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>