fix: install Ubuntu WSL distro in startup script
PR #441 added a firewall rule to allow inbound traffic to the WSL VM.
However, no WSL VM is started in this startup script, so no interfaces
were available to provide to the rule. CI changes in the main Finch repo
were taking care of launching a Ubuntu distro. Better practice would be
to install that distro in the init script of the runner itself.

Signed-off-by: Gavin Inglis <giinglis@amazon.com>